backend/requirements.txt
fastapi[standard]~=0.115.8
python-dotenv~=1.0.1
langchain~=0.3.19
langchain-community
openai
langchain-openai~=0.3.6
langchain-google-genai~=2.1.10
pymongo~=4.11.1
transformers
torch
pydantic~=2.10.6
langchain-core~=0.3.37
tqdm
weaviate-client